refactor: 分离AgentProfileCache至编排层
