+++ src/hivememory/alice/runtime/agent/profile_resolver.py
@@ -1,16 +1,40 @@
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING
+from collections import OrderedDict
+from typing import TYPE_CHECKING, Optional
 
-from hivememory.agent_runtime.cache import AgentProfileCache
-from hivememory.core.models import AgentProfile, OMNI_DOLL_PROFILE
+from hivememory.core.models import AgentProfile, MemoryAtom, MemoryType, OMNI_DOLL_PROFILE
 from hivememory.system.contracts.routes import GlobalRoutes
 
 if TYPE_CHECKING:
     from hivememory.alice.runtime.bus import AliceBus
 
 logger = logging.getLogger(__name__)
+
+
+class AgentProfileCache:
+    """人偶图纸缓存 - 会话级 LRU 缓存。"""
+
+    def __init__(self, max_size: int = 32):
+        self._max_size = max_size
+        self._cache: OrderedDict[str, tuple[Optional[MemoryAtom], AgentProfile]] = OrderedDict()
+
+    def get(self, alias: str) -> Optional[AgentProfile]:
+        entry = self._cache.get(alias)
+        if entry is not None:
+            self._cache.move_to_end(alias)
+            return entry[1]
+        return None
+
+    def store(self, alias: str, atom: Optional[MemoryAtom], config: AgentProfile) -> None:
+        if alias in self._cache:
+            self._cache.move_to_end(alias)
+            self._cache[alias] = (atom, config)
+        else:
+            if len(self._cache) >= self._max_size:
+                self._cache.popitem(last=False)
+            self._cache[alias] = (atom, config)
 
 
 class AgentProfileResolver:
@@ -39,11 +63,7 @@ class AgentProfileResolver:
 
         if profile is not None:
             logger.info(f"Agent profile '{agent_alias}' loaded and cached.")
-            self._cache.store(
-                agent_alias,
-                None,
-                profile,
-            )
+            self._cache.store(agent_alias, None, profile)
             return profile
 
         logger.info(
@@ -52,6 +72,5 @@ class AgentProfileResolver:
         return OMNI_DOLL_PROFILE
 
 
-__all__ = [
-    "AgentProfileResolver",
-]
+__all__ = ["AgentProfileResolver"]
+
